Add ability to remove downloaded CLI and PATH entry
Introduces functions to remove the downloaded CLI binary and its directory, as well as to remove the CLI directory from the user's PATH on supported platforms. Updates the Electron preload and Vue app to expose and use this functionality, including a new 'Remove Downloaded CLI' button in the UI. Also adds a 'Show alias (vx)' button to display CLI alias information.